Enhance Triangle and fix issues
 - Implement Triangle intersect
 - Fix memory leaks caused by smart pointer initialization
   in parameter list
 - Fix Vector cross method
 - Refactor minor code smells

+++ src/geometry/Plane.cpp
@@ -6,14 +6,13 @@ Plane::Plane(const Vector3 &point, const Vector3 &normal, LightIntensity color) 
 }
 
 bool Plane::intersect(const Ray &ray, std::vector<Vector3> &intersections) {
-    float denominator = this->normal.dot(ray.direction);
+    double denominator = this->normal.dot(ray.direction);
     if(denominator > 0) {
         Vector3 distance = this->point - ray.origin;
-        float t = distance.dot(this->normal) / denominator;
+        double t = distance.dot(this->normal) / denominator;
 
         if(t > 0) {
             intersections.push_back(t * ray.direction + ray.origin);
-
             return true;
         }
     }
